Fix bug with EditQuestion
The EditQuestion component was seeing this weird bug where when
you clicked the Add Answer button before the question was saved,
it would create a new blank answer (without saving to database),
and then the answer would disapear when the component updated.
It turns out the problem was that the component was getting
unmounted and recreated because of some code in the EditQuiz
component that was meant to help with deletion of a Question.
The fix was to find another solution for question deletion
and remove the offending code.